Change ssplit()'s fourth parameter from int to the proper size_t
As pointed out by clang, size_t's don't necessarily have the
same width as int's and while I'm not aware of any actual
issues caused by this (here), it doesn't hurt to use the right
type anyway.

+++ ssplit.c
@@ -74,7 +74,7 @@ const char ssplit_h_rcs[] = SSPLIT_H_VERSION;
  *                On error, vec and str may still have been overwritten.
  *
  *********************************************************************/
-int ssplit(char *str, const char *delim, char *vec[], int vec_len, 
+int ssplit(char *str, const char *delim, char *vec[], size_t vec_len,
            int dont_save_empty_fields, int ignore_leading)
 {
    unsigned char is_delim[256];
